finalizados alguns ajustes na camada de service e implementadaas execessões específicas para a camada de serviço

+++ src/main/java/com/devsuperior/clientes/services/ClientService.java
@@ -3,10 +3,15 @@ package com.devsuperior.clientes.services;
 import com.devsuperior.clientes.dto.ClientDTO;
 import com.devsuperior.clientes.entities.Client;
 import com.devsuperior.clientes.repositories.ClientRepository;
+import com.devsuperior.clientes.services.exceptions.DatabaseException;
+import com.devsuperior.clientes.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,7 +22,7 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id) {
-        Client client = clientRepository.findById(id).get();
+        Client client = clientRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Client not found!"));
         return new ClientDTO(client);
     }
 
@@ -30,29 +35,42 @@ public class ClientService {
     @Transactional
     public ClientDTO insert(ClientDTO dto) {
         Client client = new Client();
-        client.setName(dto.getName());
-        client.setCpf(dto.getCpf());
-        client.setIncome(dto.getIncome());
-        client.setBirthDate(dto.getBirthDate());
-        client.setChildren(dto.getChildren());
+        copyDtoToEntity(dto, client);
         client = clientRepository.save(client);
         return new ClientDTO(client);
     }
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
-        Client client = clientRepository.findById(id).get();
-        client.setName(dto.getName());
-        client.setCpf(dto.getCpf());
-        client.setIncome(dto.getIncome());
-        client.setBirthDate(dto.getBirthDate());
-        client.setChildren(dto.getChildren());
-        client = clientRepository.save(client);
-        return new ClientDTO(client);
+        try {
+            Client client = clientRepository.findById(id).get();
+            copyDtoToEntity(dto, client);
+            client = clientRepository.save(client);
+            return new ClientDTO(client);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Client not found!");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        clientRepository.deleteById(id);
+        if (!clientRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Client not found!");
+        }
+
+        try {
+            clientRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Integrity violation!");
+        }
     }
+
+    private void copyDtoToEntity(ClientDTO dto, Client entity) {
+        entity.setName(dto.getName());
+        entity.setCpf(dto.getCpf());
+        entity.setIncome(dto.getIncome());
+        entity.setBirthDate(dto.getBirthDate());
+        entity.setChildren(dto.getChildren());
+    }
+
 }
